backend/requirements.txt
Flask
Flask-CORS
Pillow
torch
torchvision
timm
opencv-python-headless
numpy==1.26.4